Fix workshop, improved rag, color added to the font.
